requirements.txt
Pillow>=10.0.0
matplotlib>=3.7.0
streamlit>=1.28.0
cryptography>=41.0.0
